move memory allocators to allocator.{h,cc}
