+++ caffe2/core/context_gpu.h
@@ -26,7 +26,6 @@ enum class CudaMemoryPoolType {
  */
 CudaMemoryPoolType GetCudaMemoryPoolType();
 
-
 /**
  * A struct to host thread-local cuda objects.
  *
